requirements.txt
faiss-cpu
scikit-learn
numpy
tqdm
joblib
python-dotenv
langgraph
langchain-core
langchain_mistralai